safety-tool-lib: test Memo(memo = "...") rather than Unwrap
