+++ demo/rustc_driver/safety-tool-lib/tests/snapshots/testcase_memo_property.rs
