+++ demo/rustc_driver/safety-tool-lib/tests/testcase_property.rs
@@ -6,10 +6,13 @@ use safety_tool_lib::safety;
 // #[Property(args)] syntax
 
 #[safety::precond::Align(p, T)]
-pub fn api1() {}
+pub fn align() {}
+
+#[safety::precond::Unwrap(p, T)]
+pub fn unwrap() {}
 
 #[safety::hazard::Alias(p1)]
-pub fn api2() {}
+pub fn alias() {}
 
 #[safety::option::Unreachable()]
-pub fn api3() {}
+pub fn unreachable() {}
